Rename the js components folder
